Permission 2 use Server Side Pub License elsewhere
If someone want to use this license as a template for a non-mongo app is Mongo okay with the license being used in that context? (i.e. like how people choose Apache or MIT licenses)I realize that this new licenses is trying to be open while protecting Mongo and thought that if more people used it ti could help make the license more popular. Are there any plans to release it under some creative commons or similar license. In other words, If I used the template as is but swapped Mongo's name for another software name that was totally unrelated to what the Mongo software does, would that be okay or require permission?